Started work on redoing group tasks to parallel tasks
Group tasks are being removed - I decided that tasks should be allowed
to link directly to another .toml file so groups otherwise proved
useless.
Instead I'm replacing it with a ParallelTask block, so now a set of
branches can be designed to run concurrently.

The ParallelTasks should be given a set of starting node UIDs, and a
list of end UIDs (optional).
If no end uid is provided on each branch then the concurrency will
continue until a leaf node is reached.
Split branches spawn 2 new threads per branch split.