Started refactoring internal/task/
TaskManager was becoming complex so it's getting split into
TaskEngine.go and DAGResolver.go. Now the DAGResolver should handle any
of the DAG logic, and the TaskEngine should basically only be
responsible for executing tasks in the correct order.

The DAGResolver doesn't care about the execution order, only the
structure of the DAG, and the TaskEngine shouldn't care about structure,
only the order of execution and managing that execution